fixed sonarcloud bugs

+++ NaturalAPI_Develop/API/Custumer/WithdrawCash_0.java
@@ -1,6 +1,6 @@
 public class WithdrawCash {
 
-	public double withdrawCash (double cash){
+	public int withdrawCash (double cash){
 
 	}
 }
